// src/app/domain/credit-retail/model/applicant-credit-score-stipulation-model.ts
/**
 * miles-credit-retail
 * This functional service supports the creation and follow-up of **credit applications**.    For more information, check the **Read Me** tab.
 *
 * The version of the OpenAPI document: 1.29.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EnumValue } from './enum-value';
import { FindStipulationDocument } from './find-stipulation-document';


/**
 * A stipulation linked to the applicant credit score
 */
export interface ApplicantCreditScoreStipulationModel { 
    /**
     * The ID of the stipulation
     */
    stipulationId?: string;
    stipulationStatus?: EnumValue;
    stipulationType?: EnumValue;
    /**
     * The name of the stipulation
     */
    name?: string;
    /**
     * The list of documents attached to the stipulation
     */
    documents?: Array<FindStipulationDocument>;
}